chore: Update retry group names (#5)
Drop the machine name and retry in the group name. It has been causing issues and is a bit confusing to see on the dashboard. We don't really care what machine id something is run on or how many retries there are. We only care about which retry number we're currently in